+++ CHANGELOG.md
@@ -6,8 +6,8 @@
 
 ### Features 🎉
 
-- Manual tag support added. When this tag is used build,
-  run and test will not be done without explicitly asked.
+- Manual tag and build_manual_targets support. 
+  Using this flag it is possible to build targets marked as manual
   | [#236](https://github.com/JetBrains/bazel-bsp/pull/236)
 - Build flags are reloaded during sync.
   | [#242](https://github.com/JetBrains/bazel-bsp/pull/242)